fix: work around hard crash in neovim
I could never figure out exactly why this was happening, but when the
tabpage display was left-aligned and we had enough buffers open to
require scrolling, scrolling through those buffers would lead to a
"realloc: invalid next size" stemming from buffer.c in neovim. Could not
figure out the root issue, but did find a shitty workaround that seems
to fix the crashes, so I guess we just sweep it under the rug for now.