feat: Remove animations and change rendering logic
I was getting some really crazy rendering inconsistencies, so I
simplified the rendering logic. We're now setting the tabline to call a
static lua function. A side effect of this is that I had to rip out the
animations as well, but I wanted to do that anyway. I expect there are
going to be some edges cases that spring up, so I'm going to let this
sit for a few days while I hunt down bugs.